+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Fix shape.Diamond position, xy now reflects top, left corner
 - Replace ActivityDiagram.If with Or when switching decision
 - Replace ActivityDiagram.Then with Trans and add TransRight methods
 - Activity diagram default spacing increased from 40 to 60